Added AuthGuard Added cookie auth Added Gallery Admin Panel

+++ server.js
@@ -52,12 +52,16 @@ app.get('/api/reservations', function (req, res, next) {
 });
 
 app.post('/api/reservations', function (req, res, next) {
+  if (req.cookies.auth_token !== process.env.AUTH_TOKEN) {
+    res.status(503).send({
+      status: 'User not authorized'
+    });
+    return;
+  }
   const data = {
     DateFrom : req.body.DateFrom,
     DateTo : req.body.DateTo
   }
-  console.log(req.body.DateTo);
- console.log(req.body.DateFrom);
 
  console.log(req.body);
   pool.connect((err, client, done) => {
@@ -83,6 +87,12 @@ app.delete('/api/reservations/:id', function (req, res) {
   const id = req.params.id;
 
   pool.connect((err, client, done) => {
+    if (req.cookies.auth_token !== process.env.AUTH_TOKEN) {
+      res.status(503).send({
+        status: 'User not authorized'
+      });
+      return;
+    }
     const query = 'DELETE FROM "dbo.Reservations" WHERE "Id" = $1';
     const values = [id];
 
@@ -98,25 +108,111 @@ app.delete('/api/reservations/:id', function (req, res) {
   });
 });
 
+app.get('/api/gallery', function (req, res, next) {
+  
+  pool.connect(function(err, client, done) {
+      if(err){
+          console.log(err);
+          throw err;
+      } 
+      client.query(`SELECT "Id", "Url" FROM "dbo.Gallery" ORDER BY "Id" DESC`, function(err, result) {
+          done();
+          if(err) {
+              res.status(400).send(err);
+              console.log(err);
+              return;
+          }
+          res.status(200).send(result.rows);
+      });
+  });
+});
 
-app.get('/api/login', function (req, res, next) {
-  console.log('api login');
-  // const credentials = {
-  //   login: req.body.user,
-  //   pass: req.body.pass
-  // }
-
-  // if (credentials.login === process.env.ADMIN_USER &&
-  //     credentials.pass === process.env.ADMIN_PASS) {
-
-  //     } 
-  //console.log(credentials);
-
-  if (req.cookies) {
-    console.log(req.cookies.loginName);
+app.post('/api/gallery', function (req, res, next) {
+  if (req.cookies.auth_token !== process.env.AUTH_TOKEN) {
+    res.status(503).send({
+      status: 'User not authorized'
+    });
+    return;
   }
-  res.cookie('loginName',"John Doe", { maxAge: 20 *60 * 1000, httpOnly: true });
-  res.send("test");
+
+  const data = {
+    Url : req.body.Url
+  }
+  
+  pool.connect((err, client, done) => {
+    if(err){
+      console.log(err);
+      throw err;
+    }
+    const query = 'INSERT INTO "dbo.Gallery"("Url") VALUES ($1) RETURNING "Id"';
+    const values = [data.Url];
+    client.query(query, values, (error, result) => {
+      done();
+      if (error) {
+        res.status(400).json({error});
+        return;
+      }
+      res.status(202).send(result.rows[0]);
+    });
+  });
+});
+
+app.delete('/api/gallery/:id', function (req, res) {
+  
+  const id = req.params.id;
+  if (req.cookies.auth_token !== process.env.AUTH_TOKEN) {
+    res.status(503).send({
+      status: 'User not authorized'
+    });
+    return;
+  }
+
+  pool.connect((err, client, done) => {
+    if(err){
+      console.log(err);
+      throw err;
+    } 
+    const query = 'DELETE FROM "dbo.Gallery" WHERE "Id" = $1';
+    const values = [id];
+
+    client.query(query, values, (error, result) => {
+      done();
+      if (error) {
+        res.status(400).json({error});
+      }
+      res.status(202).send({
+        status: 'Successful'
+      });
+    });
+  });
+});
+
+app.post('/api/validate', function (req, res, next) {
+  if (req.cookies.auth_token === process.env.AUTH_TOKEN) {
+    res.send(true);
+  }
+  else {
+    res.send(false);
+  }
+});
+
+app.post('/api/login', function (req, res, next) {
+  console.log(req.cookies.auth_token);
+  console.log('api login');
+  const credentials = {
+    login: req.body.user,
+    pass: req.body.pass
+  }
+
+  if (credentials.login === process.env.ADMIN_USER &&
+      credentials.pass === process.env.ADMIN_PASS) {
+        res.cookie('auth_token', process.env.AUTH_TOKEN, { maxAge: 20 *60 * 1000, httpOnly: true, sameSite: 'strict' });
+        res.send({ loggedIn: true });
+  }
+  else {
+    res.clearCookie("auth_token");
+    res.send({ loggedIn: false });
+  }
 });
 
 app.listen(process.env.PORT || 8080);
